Extract JsonRpcClientBuilder
This is to be consistent with JsonRpcServerBuilder
