+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/server/JsonRpcServer.scala
@@ -1,5 +1,6 @@
 package io.github.shogowada.scala.jsonrpc.server
 
+import io.github.shogowada.scala.jsonrpc.Models.JsonRpcMethod
 import io.github.shogowada.scala.jsonrpc.serializers.JsonSerializer
 import io.github.shogowada.scala.jsonrpc.server.JsonRpcServer.Handler
 import io.github.shogowada.scala.jsonrpc.utils.MacroUtils
@@ -32,25 +33,28 @@ object JsonRpcServerMacro {
     val jsonSerializer: Tree = q"${c.prefix.tree}.jsonSerializer"
     val methodNameToHandlerMap: Tree = q"${c.prefix.tree}.methodNameToHandlerMap"
 
-    val maybeJsonRpcMethod =
+    val maybeJsonRpcMethod = c.Expr[Option[JsonRpcMethod]](
       q"""
           $jsonSerializer.deserialize[JsonRpcMethod]($json)
               .filter(method => method.jsonrpc == Constants.JsonRpc)
           """
+    )
 
-    val maybeHandler =
+    val maybeHandler = c.Expr[Option[Handler]](
       q"""
           $maybeJsonRpcMethod
               .flatMap((jsonRpcMethod: JsonRpcMethod) => {
                 $methodNameToHandlerMap.get(jsonRpcMethod.method)
               })
           """
+    )
 
-    val futureMaybeJson =
+    val futureMaybeJson = c.Expr[Future[Option[String]]](
       q"""
           $maybeHandler.map(handler => handler($json))
             .getOrElse(Future(None))
           """
+    )
 
     c.Expr(
       q"""
