+++ shared/src/test/scala/io/github/shogowada/scala/jsonrpc/client/JsonRpcClientTest.scala
@@ -10,7 +10,8 @@ class JsonRpcClientTest extends path.FunSpec
   override def newInstance: path.FunSpecLike = new JsonRpcClientTest
 
   implicit val executionContext = ExecutionContext.Implicits.global
-  val target = JsonRpcClient(UpickleJsonSerializer(), (json: String) => {})
+  val target = JsonRpcClientBuilder(UpickleJsonSerializer(), (json: String) => {})
+      .build
 
   describe("given I have an API") {
     trait Api {
